CIF-991 - MagentoGraphqlClient & Utils should be public
