+++ bundles/core/pom.xml
@@ -496,7 +496,7 @@
         <dependency>
             <groupId>com.adobe.commerce.cif</groupId>
             <artifactId>magento-graphql</artifactId>
-            <version>3.0.0-magento232</version>
+            <version>3.0.1-magento232</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
