+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/internal/models/v1/categorylist/FeaturedCateogoryListImpl.java
@@ -31,9 +31,9 @@ import org.apache.sling.models.annotations.injectorspecific.ScriptVariable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.adobe.cq.commerce.core.components.internal.models.v1.MagentoGraphqlClient;
-import com.adobe.cq.commerce.core.components.internal.models.v1.Utils;
+import com.adobe.cq.commerce.core.components.client.MagentoGraphqlClient;
 import com.adobe.cq.commerce.core.components.models.categorylist.FeaturedCategoryList;
+import com.adobe.cq.commerce.core.components.utils.SiteNavigation;
 import com.adobe.cq.commerce.graphql.client.GraphqlResponse;
 import com.adobe.cq.commerce.magento.graphql.*;
 import com.adobe.cq.commerce.magento.graphql.gson.Error;
@@ -66,7 +66,7 @@ public class FeaturedCateogoryListImpl implements FeaturedCategoryList {
 
     @PostConstruct
     private void initModel() {
-        categoryPage = Utils.getCategoryPage(currentPage);
+        categoryPage = SiteNavigation.getCategoryPage(currentPage);
         if (categoryPage == null) {
             categoryPage = currentPage;
         }
